+++ src/main/kotlin/sogong/sogongSpring/repository/HashtagDbRepository.kt
@@ -3,5 +3,5 @@ package sogong.sogongSpring.repository
 import org.springframework.data.jpa.repository.JpaRepository
 import sogong.sogongSpring.entity.HashtagDbEntity
 
-interface HashtagDbRepository : JpaRepository<HashtagDbEntity, Int> {
+interface HashtagDbRepository : JpaRepository<HashtagDbEntity, Long> {
 }
